Issue #132 refactor: 리스트 어댑터 diffUtil 위치 변경
diffUtil을 재사용할 필요가 있어 diffUtil의 위치를 아이템 클래스의 companion
object로 변경함.
