+++ app/src/main/java/com/drunkenboys/calendarun/ui/searchschedule/SearchScheduleAdapter.kt
@@ -2,7 +2,6 @@ package com.drunkenboys.calendarun.ui.searchschedule
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.drunkenboys.calendarun.BR
 import com.drunkenboys.calendarun.R
@@ -11,7 +10,7 @@ import com.drunkenboys.calendarun.databinding.ItemDateScheduleBinding
 import com.drunkenboys.calendarun.ui.base.BaseViewHolder
 import com.drunkenboys.calendarun.ui.searchschedule.model.DateItem
 
-class SearchScheduleAdapter : ListAdapter<DateItem, BaseViewHolder<ItemDateBinding>>(diffUtil) {
+class SearchScheduleAdapter : ListAdapter<DateItem, BaseViewHolder<ItemDateBinding>>(DateItem.diffUtil) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): BaseViewHolder<ItemDateBinding> =
         BaseViewHolder(parent, R.layout.item_date)
@@ -28,16 +27,6 @@ class SearchScheduleAdapter : ListAdapter<DateItem, BaseViewHolder<ItemDateBindi
             val dateScheduleBinding = ItemDateScheduleBinding.inflate(LayoutInflater.from(root.context), layoutDateSchedule, false)
             dateScheduleBinding.item = schedule
             layoutDateSchedule.addView(dateScheduleBinding.root)
-        }
-    }
-
-    companion object {
-
-        private val diffUtil = object : DiffUtil.ItemCallback<DateItem>() {
-
-            override fun areItemsTheSame(oldItem: DateItem, newItem: DateItem) = oldItem.hashCode() == newItem.hashCode()
-
-            override fun areContentsTheSame(oldItem: DateItem, newItem: DateItem) = oldItem == newItem
         }
     }
 }
