+++ app/src/main/java/com/drunkenboys/calendarun/ui/searchschedule/model/DateScheduleItem.kt
@@ -1,5 +1,6 @@
 package com.drunkenboys.calendarun.ui.searchschedule.model
 
+import androidx.recyclerview.widget.DiffUtil
 import com.drunkenboys.calendarun.data.schedule.entity.Schedule
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
@@ -19,6 +20,16 @@ data class DateScheduleItem(val schedule: Schedule, val onClick: () -> Unit) {
             startDate.month < endDate.month -> DateTimeFormatter.ofPattern("M월 d일 hh:mm")
             startDate.dayOfYear < endDate.dayOfYear -> DateTimeFormatter.ofPattern("d일 hh:mm")
             else -> DateTimeFormatter.ofPattern("hh:mm")
+        }
+    }
+
+    companion object {
+
+        val diffUtil = object : DiffUtil.ItemCallback<DateScheduleItem>() {
+            override fun areItemsTheSame(oldItem: DateScheduleItem, newItem: DateScheduleItem) =
+                oldItem.schedule.id == newItem.schedule.id
+
+            override fun areContentsTheSame(oldItem: DateScheduleItem, newItem: DateScheduleItem) = oldItem == newItem
         }
     }
 }
